Extract v3 asset downloading in separate request strategy
